Rewrite buildSearchQuery
