+++ src/modules/Datasets/components/Datasets/__test__/Datasets.test.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { shallow, mount } from 'enzyme'
+import { buildSearchQuery } from '../Datasets'
 
 const Datasets = require('proxyquire')('../Datasets', {
   '../../../../fetch/fetch': require('../../../../../fetch/__mocks__/fetch'),
@@ -7,6 +8,28 @@ const Datasets = require('proxyquire')('../Datasets', {
     browserHistory: {push:  () => {}}
   }
 }).default
+
+describe('buildSearchQuery()', () => {
+  it('should return a query', () => {
+    const componentState = {
+      textInput: '42',
+      page: 2,
+      filters: [
+        {name: 'keywords', value: 'keyword1'},
+        {name: 'keywords', value: 'keyword2'},
+        {name: 'organizations', value: 'foo'},
+      ],
+    }
+    const expectedUrl = 'q=42&page=2&keywords=keyword1&keywords=keyword2&organizations=foo'
+
+    const builQuery = buildSearchQuery(
+      componentState.textInput,
+      componentState.filters,
+      componentState.page,
+    )
+    expect(builQuery).to.equal(expectedUrl)
+  })
+})
 
 describe('<Datasets />', () => {
